# Use strings.Builder and strconv.Quote