--- a/README.md
+++ b/README.md
@@ -55,7 +55,7 @@ func getHTTPRequestCounter(host string) *vm.Counter {
 }
 ```
 
-Vimebu also exposes a way to escape quotes on label values you don't control using `Builder.LabelAppendQuote`.
+Vimebu also exposes a way to escape quotes on label values you don't control using `Builder.LabelQuote`.
 ```go
 import (
     "github.com/wazazaby/vimebu"
@@ -65,7 +65,7 @@ import (
 func getHTTPRequestCounter(path string) *vm.Counter {
     metric := vimebu.
         Metric("api_http_requests_total").
-        LabelAppendQuote("path", path).
+        LabelQuote("path", path).
         String() // api_http_requests_total{path="some/bro\"ken/path"}
     return vm.GetOrCreateCounter(metric)
 }