At first, I thought that using `bytes.Buffer` + `strconv.AppendQuote` was faster than using `strings.Builder` + `strconv.Quote`, but after some benchmarking it turns out it's not the case. Here are some samples.

<details>
<summary>Using bytes.Buffer</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/wazazaby/vimebu
BenchmarkBuilder-10                   	 3414580	       338.0 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilder-10                   	 3540535	       339.2 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilder-10                   	 3522084	       345.2 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilder-10                   	 3340707	       350.1 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilder-10                   	 3545394	       337.7 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilder-10                   	 3552984	       338.9 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilder-10                   	 3550158	       349.3 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilder-10                   	 3544591	       339.0 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilder-10                   	 3552189	       346.2 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilder-10                   	 3537568	       340.9 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8593497	       142.8 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8372839	       139.8 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8690164	       143.2 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8790312	       138.8 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8814727	       145.3 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8683002	       139.1 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8632513	       137.0 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8700102	       137.3 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8453677	       142.9 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	 8058141	       141.8 ns/op	     320 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2281246	       536.1 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2235252	       529.9 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2266074	       537.5 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2079572	       546.3 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2272840	       523.4 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2296305	       517.3 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2266734	       528.8 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2303331	       513.5 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2317713	       518.1 ns/op	     416 B/op	       5 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2313747	       508.2 ns/op	     416 B/op	       5 allocs/op
PASS
ok  	github.com/wazazaby/vimebu	92.885s
```
</details>

<details>
<summary>Using strings.Builder</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/wazazaby/vimebu
BenchmarkBuilder-10                   	 3936693	       289.7 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilder-10                   	 4173040	       290.9 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilder-10                   	 4153046	       289.5 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilder-10                   	 4162498	       289.7 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilder-10                   	 4099214	       289.8 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilder-10                   	 3807670	       293.9 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilder-10                   	 4130784	       292.3 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilder-10                   	 4075612	       290.0 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilder-10                   	 4167272	       293.4 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilder-10                   	 4147341	       289.4 ns/op	     272 B/op	       4 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11363053	       105.4 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11381712	       105.7 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11299815	       105.4 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11400986	       105.2 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11421717	       106.0 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11333067	       105.3 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11269770	       106.0 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11386022	       106.2 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11489625	       105.6 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteNone-10    	11416881	       105.6 ns/op	     224 B/op	       3 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2322031	       517.3 ns/op	     312 B/op	       7 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2321367	       518.2 ns/op	     312 B/op	       7 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2313603	       517.6 ns/op	     312 B/op	       7 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2324313	       516.8 ns/op	     312 B/op	       7 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2323504	       518.4 ns/op	     312 B/op	       7 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2313649	       531.0 ns/op	     312 B/op	       7 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2303410	       526.9 ns/op	     312 B/op	       7 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2311272	       517.1 ns/op	     312 B/op	       7 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2283028	       528.0 ns/op	     312 B/op	       7 allocs/op
BenchmarkBuilderAppendQuoteOnly-10    	 2290184	       847.7 ns/op	     312 B/op	       7 allocs/op
PASS
ok  	github.com/wazazaby/vimebu	92.718s
```
</details>

Comparing the two with `benchstat` gives me this result :
```
goos: darwin
goarch: arm64
pkg: github.com/wazazaby/vimebu
                          │ bytes.buffer.txt │         strings.builder.txt         │
                          │      sec/op      │   sec/op     vs base                │
Builder-10                       341.6n ± 1%   289.8n ± 0%  -15.16% (p=0.000 n=20)
BuilderAppendQuoteNone-10        139.5n ± 2%   105.7n ± 0%  -24.27% (p=0.000 n=20)
BuilderAppendQuoteOnly-10        517.6n ± 2%   522.0n ± 1%        ~ (p=0.086 n=20)
geomean                          291.1n        251.9n       -13.47%

                          │ bytes.buffer.txt │        strings.builder.txt         │
                          │       B/op       │    B/op     vs base                │
Builder-10                        416.0 ± 0%   272.0 ± 0%  -34.62% (p=0.000 n=20)
BuilderAppendQuoteNone-10         320.0 ± 0%   224.0 ± 0%  -30.00% (p=0.000 n=20)
BuilderAppendQuoteOnly-10         416.0 ± 0%   312.0 ± 0%  -25.00% (p=0.000 n=20)
geomean                           381.2        266.9       -29.98%

                          │ bytes.buffer.txt │         strings.builder.txt          │
                          │    allocs/op     │ allocs/op   vs base                  │
Builder-10                        5.000 ± 0%   4.000 ± 0%  -20.00% (p=0.000 n=20)
BuilderAppendQuoteNone-10         3.000 ± 0%   3.000 ± 0%        ~ (p=1.000 n=20) ¹
BuilderAppendQuoteOnly-10         5.000 ± 0%   7.000 ± 0%  +40.00% (p=0.000 n=20)
geomean                           4.217        4.380        +3.85%
¹ all samples are equal
```

As you can see, using `strings.Builder` is about ~13% faster and it allocates about ~30% less bytes. The only downside is that is makes 2 more allocations per operation when every label value needs to be quoted.